requirements.txt
Pillow
ultralytics
opencv-python
